all done

+++ src/Pages/Admin/Admin.jsx
@@ -23,7 +23,6 @@ const Admin = () => {
             });
     };
     // fetching all bookings
-
     // fetching data  here
     // const [allBookings, setAllBookings] = useState([]);
     useEffect(() => {
@@ -36,19 +35,22 @@ const Admin = () => {
     }, []);
     //delete single booking
     const handleDelete = (id) => {
-        fetch(`https://gruesome-beast-12739.herokuapp.com/booking/${id}`, {
-            method: "DELETE",
-        })
-            .then((res) => res.json())
-            .then((data) => {
-                if (data.deletedCount > 0) {
-                    alert("Canceled Successfully.");
-                    const remainingBookings = allBookings.filter(
-                        (booking) => booking._id !== id
-                    );
-                    setAllBookings(remainingBookings);
-                }
-            });
+        const proceed = window.confirm("Are you sure, you want to cancel?");
+        if (proceed) {
+            fetch(`https://gruesome-beast-12739.herokuapp.com/booking/${id}`, {
+                method: "DELETE",
+            })
+                .then((res) => res.json())
+                .then((data) => {
+                    if (data.deletedCount > 0) {
+                        alert("Canceled Successfully.");
+                        const remainingBookings = allBookings.filter(
+                            (booking) => booking._id !== id
+                        );
+                        setAllBookings(remainingBookings);
+                    }
+                });
+        }
     };
     // single data
     const [hotel, setHotel] = useState({});
@@ -137,29 +139,32 @@ const Admin = () => {
                                     <div className="row">
                                         <div className="col-sm-12">
                                             {allBookings.map((booking) => (
-                                                <div class="card m-4">
+                                                <div
+                                                    key={booking?._id}
+                                                    className="card m-4"
+                                                >
                                                     <img
                                                         src={booking?.image}
-                                                        class="card-img-top"
+                                                        className="card-img-top"
                                                         alt="..."
                                                     />
                                                     <div className="card-body">
-                                                        <h5 class="card-title">
+                                                        <h5 className="card-title">
                                                             {booking?.hotelname}
                                                         </h5>
-                                                        <p class="card-text">
+                                                        <p className="card-text">
                                                             Booked By :{" "}
                                                             <b>
                                                                 {booking?.name}
                                                             </b>
                                                         </p>
-                                                        <p class="card-text">
+                                                        <p className="card-text">
                                                             Email :{" "}
                                                             <b>
                                                                 {booking?.email}
                                                             </b>
                                                         </p>
-                                                        <p class="card-text">
+                                                        <p className="card-text">
                                                             Status :{" "}
                                                             {booking?.status}
                                                         </p>
@@ -205,7 +210,7 @@ const Admin = () => {
                                     </thead>
                                     <tbody>
                                         {allBookings.map((booking) => (
-                                            <tr>
+                                            <tr key={booking?._id}>
                                                 <td>{booking?.name}</td>
                                                 <td>{booking?.email}</td>
                                                 <td>20/07/2021</td>
